% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusteringClass.R
\docType{class}
\name{ClusteringClass-class}
\alias{ClusteringClass-class}
\title{ClusteringClass}
\description{
A class for clustering-related tasks
}
\section{Slots}{

\describe{
\item{\code{runID}}{A unique identifier for each clustering instance.}

\item{\code{method}}{The name of the clustering method used.}

\item{\code{result}}{The clustering result in matrix format.}

\item{\code{methodSpecificResults}}{A list to store method-specific outputs.}
}}

