% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoCrunch.R
\docType{class}
\name{CytoCrunch-class}
\alias{CytoCrunch-class}
\alias{CytoCrunch}
\title{The CytoCrunch Class}
\format{
An object of class \code{CytoCrunch} is a list of slots that store information about the sample, expression data, and results of clustering and dimensionality reduction methods.
Initialize an empty plotList when creating a new CytoCrunch object
}
\usage{
\S4method{initialize}{CytoCrunch}(.Object, sampleName, inputCSVFilePath, ...)
}
\description{
The CytoCrunch object is a representation of a gated population of cells that
are exported from a FlowJo plugin process. These files loaded into the object
can be either .fcs files or .csv files containing expression data from a flow
cytometery experiment. The object stores metadata as well as representations
of low-dimensional data from clustering or dimensionality reductions methods.

Each CytoCrunch object has a number of slots which store information. Key
slots to access are listed below.
}
\section{Slots}{

\describe{
\item{\code{sampleName}}{The unique sample name for the selected file}

\item{\code{inputCSV}}{The expression matrix exported from FlowJo}

\item{\code{parameterNames}}{The metadata for parameter and stain names}

\item{\code{flowFrameFJ}}{The expression matrix stored as flowFrame}

\item{\code{clusteringList}}{A list of ClusteringClass objects}

\item{\code{dimReductionList}}{A list of DimRedClass objects}

\item{\code{batchCorrectionList}}{A list of BatchCorrectionClass objects}
}}

