% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoCrunchMerged.R
\docType{class}
\name{CytoCrunchMerged-class}
\alias{CytoCrunchMerged-class}
\title{CytoCrunchMerged}
\description{
A class for storing merged CytoCrunch objects
}
\section{Slots}{

\describe{
\item{\code{sampleNames}}{A character vector of sample names.}

\item{\code{runIDs}}{A character vector of run IDs.}

\item{\code{embeddingTypes}}{A character vector of embedding types.}

\item{\code{mergedEmbeddings}}{A list of merged embeddings.}

\item{\code{mergedClusters}}{A list of merged clusters.}

\item{\code{additionalData}}{A list of additional data.}
}}

\examples{
# Create a CytoCrunchMerged object
mergedObject <- new("CytoCrunchMerged",
                   sampleNames = c("Sample1", "Sample2"),
                   runIDs = c("Run1", "Run2"),
                   embeddingTypes = c("tSNE", "UMAP"),
                   mergedEmbeddings = list(matrix(NA, 5, 2), matrix(NA, 5, 2)),
                   mergedClusters = list(factor(1:5), factor(1:5)),
                   additionalData = list(list(), list())
                   )
# Print the CytoCrunchMerged object
print(mergedObject)
}
