% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DimRedClass.R
\docType{class}
\name{DimRedClass-class}
\alias{DimRedClass-class}
\title{DimRedClass}
\description{
A class for dimensionality reduction-related tasks
}
\section{Slots}{

\describe{
\item{\code{runID}}{A unique identifier for each dim reduction instance.}

\item{\code{method}}{The name of the dim reduction method used.}

\item{\code{result}}{The dim reduction result in matrix format.}

\item{\code{methodSpecificResults}}{A list of method-specific results.}
}}

