% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoCrunch.R
\docType{methods}
\name{getParNames}
\alias{getParNames}
\alias{getParNames,CytoCrunch-method}
\title{View parameter names}
\usage{
getParNames(object)

\S4method{getParNames}{CytoCrunch}(object)
}
\arguments{
\item{object}{A CytoCrunch object}
}
\value{
Returns the parameter names from the exported population
}
\description{
This function returns the parameter names if present from the exported population
}
\examples{
\dontrun{
getParNames(object)}
}
