% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoCrunch.R
\docType{methods}
\name{getSampleName}
\alias{getSampleName}
\alias{getSampleName,CytoCrunch-method}
\title{View the sample name}
\usage{
getSampleName(object)

\S4method{getSampleName}{CytoCrunch}(object)
}
\arguments{
\item{object}{A CytoCrunch object}
}
\value{
Returns the sampleName
}
\description{
This function returns the sample name from the CytoCrunch object.
}
\examples{
\dontrun{
getSampleName(object)}
}
