% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoCrunch.R
\docType{methods}
\name{getStainNames}
\alias{getStainNames}
\alias{getStainNames,CytoCrunch-method}
\title{View stain names}
\usage{
getStainNames(object)

\S4method{getStainNames}{CytoCrunch}(object)
}
\arguments{
\item{object}{A CytoCrunch object}
}
\value{
Returns the stain names if present from the exported population
}
\description{
This function returns the stain names if present from the exported population
}
\examples{
\dontrun{
getStainNames(object)}
}
