% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoCrunch.R
\docType{methods}
\name{loadFlowJoCSV}
\alias{loadFlowJoCSV}
\alias{loadFlowJoCSV,CytoCrunch-method}
\title{Load data into CytoCrunch object}
\usage{
loadFlowJoCSV(object, csvPath, eventNumber = "EventNumberDP")

\S4method{loadFlowJoCSV}{CytoCrunch}(object, csvPath, eventNumber = "EventNumberDP")
}
\arguments{
\item{object}{A CytoCrunch object}

\item{csvPath}{The path to the exported csv file from FlowJo}

\item{eventNumber}{The column name for event numbers or cell ids}
}
\value{
Returns the newly created object
}
\description{
Load data into CytoCrunch object
}
\examples{
\dontrun{
inputCSVpath <- "/Documents/ExtNode.csv"
myobject <- loadFlowJoCSV(object, inputCSVpath, eventNumber = "EventNumberDP")}
}
