% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoCrunchMerged.R
\name{mergeCytoCrunchObjects}
\alias{mergeCytoCrunchObjects}
\title{mergeCytoCrunchObjects}
\usage{
mergeCytoCrunchObjects(objects, runIDs)
}
\arguments{
\item{objects}{A list of CytoCrunch objects to merge.}

\item{runIDs}{A character vector of run IDs corresponding to each CytoCrunch object.}
}
\value{
A CytoCrunchMerged object containing the merged data.
}
\description{
Merge multiple CytoCrunch objects into a single CytoCrunchMerged object
}
\examples{
# Create two CytoCrunch objects
obj1 <- new("CytoCrunch",
           sampleName = "Sample1",
           inputCSVFilePath = "data/sample1.csv"
           )
obj2 <- new("CytoCrunch",
          sampleName = "Sample2",
          inputCSVFilePath = "data/sample2.csv"
          )
# Merge the CytoCrunch objects
mergedObject <- mergeCytoCrunchObjects(list(obj1, obj2), c("Run1", "Run2"))
print(mergedObject)
}
