% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoCrunch.R
\docType{methods}
\name{performClustering}
\alias{performClustering}
\alias{performClustering,CytoCrunch-method}
\title{Perform clustering on the data}
\usage{
performClustering(object, runID, method = "FlowSOM", k = NULL, ...)

\S4method{performClustering}{CytoCrunch}(object, runID = "808", method = "FlowSOM", k = NULL, ...)
}
\arguments{
\item{object}{A CytoCrunch object}

\item{runID}{A unique identifier for this run}

\item{method}{The clustering method to use}

\item{k}{The number of clusters to create}

\item{...}{Additional arguments to pass to the method}
}
\value{
Returns the CytoCrunch object with the new ClusteringClass object added
}
\description{
Perform clustering on the data
}
\examples{
\dontrun{ performClustering(object, runID = "808", method = "FlowSOM", k = 10)}
}
