% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoCrunch.R
\docType{methods}
\name{performEmbedding}
\alias{performEmbedding}
\alias{performEmbedding,CytoCrunch-method}
\title{Perform dimensionality reduction with EmbedSOM}
\usage{
performEmbedding(object, runID, method = "umap", ...)

\S4method{performEmbedding}{CytoCrunch}(
  object,
  runID,
  method = "umap",
  fjSeed,
  fjXdim,
  fjYdim,
  fjRLEN,
  fjK,
  fjAdjust,
  fjSmooth,
  fjUmap.Dist,
  fjMapSize,
  supLabel,
  fjSuperWeight
)
}
\arguments{
\item{object}{A CytoCrunch object}

\item{runID}{A unique identifier for this run}

\item{method}{The dimensionality reduction method to use}

\item{...}{Additional arguments to pass to the method}
}
\value{
Returns the CytoCrunch object with the new DimRedClass object added
}
\description{
Perform dimensionality reduction with EmbedSOM
}
\examples{
\dontrun{
performEmbedding(object, runID = "808", method = "umap", fjSeed = 123,fjXdim = 10, fjYdim = 10, fjRLEN = 15, fjK = 25, fjAdjust = 0,fjSmooth = 0, fjUmap.Dist = 0.5, fjMapSize = 5000)
}
}
