% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClass.R
\docType{class}
\name{PlotClass-class}
\alias{PlotClass-class}
\title{PlotClass}
\description{
A class to store plots generated by the CytoCrunch package
}
\section{Slots}{

\describe{
\item{\code{runID}}{A character string to identify the run}

\item{\code{plotType}}{A character string to identify the type of plot}

\item{\code{plotObject}}{An object of any type that represents the plot}

\item{\code{description}}{An optional character string to store metadata or a description of the plot}
}}

\examples{
plotObj <- createPlotClass(runID = "testRun", plotType = "scatter")
plotObj
}
