% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoCrunch.R
\docType{methods}
\name{plotEmbeddingWithClusters}
\alias{plotEmbeddingWithClusters}
\title{Plot an embedding with clusters}
\format{
A ggplot2 object
}
\usage{
plotEmbeddingWithClusters(object, embeddingRunID, clusteringRunID, ...)
}
\arguments{
\item{object}{A CytoCrunch object}

\item{embeddingRunID}{The runID of the embedding to plot}

\item{clusteringRunID}{The runID of the clustering to overlay}
}
\value{
Returns a ggplot2 object
}
\description{
This function plots an embedding with clusters overlaid.
}
\examples{
\dontrun{ plotEmbeddingWithClusters(object, embeddingRunID = "SLK9", clusteringRunID = "SLK8")}
}
