% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoCrunchMerged.R
\name{plotMergedEmbeddings}
\alias{plotMergedEmbeddings}
\title{Plot Merged Embeddings}
\usage{
plotMergedEmbeddings(mergedObject, ...)
}
\arguments{
\item{mergedObject}{A CytoCrunchMerged object containing the merged embeddings}
}
\value{
A grid of plots showing the merged embeddings
}
\description{
Plot the merged embeddings from a CytoCrunchMerged object
}
\examples{
# Create a CytoCrunchMerged object
mergedObject <- new("CytoCrunchMerged",
                  sampleNames = c("Sample1", "Sample2"),
                  runIDs = c("Run1", "Run2"),
                  embeddingTypes = c("tSNE", "UMAP"),
                  mergedEmbeddings = list(matrix(NA, 5, 2), matrix(NA, 5, 2)),
                  mergedClusters = list(factor(1:5), factor(1:5)),
                  additionalData = list(list(), list())
                  )
# Plot the merged embeddings
plotMergedEmbeddings(mergedObject)
}
